Rename folder and update some syntax
