+++ packages/truffle-contract/lib/contract/bootstrap.js
