+++ packages/truffle-contract/lib/contract/constructorMethods.js
