+++ packages/truffle-contract/lib/contract/index.js
