+++ packages/truffle-contract/lib/contract/properties.js
